Hunter Ball made this
For all legalities, the entire website is intellectual property of Hunter Ball, including the logo, assets, images, and everything else. Sue me, it doesn't matter. I am the heart and soul of FractylDev, and without me you'd have a trash Morse code app that doesn't even work on Android anymore. That's it. Have fun at state, and I hope you know I can take this website down in an instant. I paid for it fully. I developed and designed it, time after time, with my own ideas and code. Nothing on this website but the namesake can be attributed to the company.